+++ src/enums/__init__.py
@@ -1,0 +1,9 @@
+from .monocratic_decision import MonocraticDecisionXPathOptions
+from .acordao import AcordaoXPathOptions
+from .duvida import DuvidaXPathOptions
+
+__all__ = [
+    MonocraticDecisionXPathOptions,
+    AcordaoXPathOptions,
+    DuvidaXPathOptions
+]
