+++ src/repositories/abstract.py
@@ -1,0 +1,6 @@
+import abc
+
+class AbstractRepository(abc.ABC):
+    @abc.abstractmethod
+    def get(self, reference):
+        raise NotImplementedError
