+++ src/scrape.py
@@ -16,11 +16,11 @@ from pagination import (
     get_next_page_element,
     get_page_information_text,
 )
-from repository import (
+from repositories.decision_details import (
     DecisionDetailsPageRepository,
-    DuvidaXPathOptions,
-    MonocraticDecisionXPathOptions,
 )
+from enums import DuvidaXPathOptions, MonocraticDecisionXPathOptions
+
 from search import fill_judgement_end, fill_judgement_start, post_search
 from transforms import convert_model_list_to_dataframe
 from validators.webpage import get_decision_type, is_decision_justice_secret
